+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This plugin is developed and maintained at https://github.com/ONLYOFFICE/onlyoff
 - Turkish and Galician empty file templates
 - jwt header setting
 - connection to the demo server
+- setting address for internal requests from the server
+- Link to docs cloud
 
 ## Changed
 - document server v6.0 and earlier is no longer supported